requirements.txt
django 
djangorestframework
couldinary
djangorestframework-simplejwt
drf-yasg